+++ bank/src/main/java/com/fisa/bank/common/config/security/auth/CustomUserDetails.java
@@ -1,4 +1,4 @@
-package com.fisa.bank.common.config.security.oauth2;
+package com.fisa.bank.common.config.security.auth;
 
 import com.fisa.bank.user.persistence.entity.id.UserId;
 import java.util.Collection;
